refactored the notes section
