+++ frontend/src/functionality/Dashboard/NotesLoading.jsx
@@ -1,0 +1,10 @@
+/**
+ * Displays a loading indicator for the notes section.
+ *
+ * @returns {React.JSX.Element} A loading message paragraph
+ */
+function NotesLoading() {
+    return <p className="notes-loading">Loading notes...</p>;
+}
+
+export default NotesLoading;
