+++ frontend/src/functionality/Dashboard/NotesTextArea.jsx
@@ -1,0 +1,20 @@
+/**
+ * Renders the notes textarea input.
+ *
+ * @param {Object} props
+ * @param {string} props.notes - The current notes value
+ * @param {Function} props.setNotes - Setter to update notes
+ * @return {React.JSX.Element} - The notes area
+ */
+function NotesTextarea({ notes, setNotes }) {
+    return (
+        <textarea
+            className="notes-textarea"
+            placeholder="Notes"
+            value={notes}
+            onChange={(e) => setNotes(e.target.value)}
+        />
+    );
+}
+
+export default NotesTextarea;
