+++ frontend/src/functionality/Dashboard/stylesheets/NotesTextArea.css
@@ -1,0 +1,24 @@
+.notes-textarea {
+    flex: 1;
+    width: 100%;
+    box-sizing: border-box;
+    resize: none;
+    padding: var(--space-sm) var(--space-navbar-x) var(--space-md) var(--space-xl);
+    background: transparent;
+    color: var(--text-main);
+    border: none;
+    border-radius: 0;
+    outline: none;
+    font-size: var(--font-size-body);
+    font-family: inherit;
+    overflow-y: auto;
+}
+
+.notes-textarea::placeholder {
+    color: var(--text-muted);
+    opacity: 1;
+}
+
+.notes-textarea:focus {
+    box-shadow: none;
+}
